requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.3.1
aiounittest==1.4.2
apturl==0.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
blinker==1.4
Brlapi==0.7.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
Cython==0.29.28
dbus-python==1.2.16
defer==1.0.6
discord==1.7.3
discord.py==1.7.3
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
editdistance==0.6.1
entrypoints==0.3
filetype==1.0.10
frozenlist==1.3.0
httplib2==0.14.0
idna==2.8
joblib==1.1.0
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
meson==0.63.2
multidict==6.0.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
scikit-learn==1.1.1
scipy==1.8.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
threadpoolctl==3.1.0
toml==0.10.0
typing-extensions==4.3.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wrapt==1.14.1
xkit==0.0.0
yarl==1.7.2
youtube-dl==2020.3.24
